Add lid-close sleep, configurable sleep buttons, and new config options
Inject a configurable GBA button combo on lid close to trigger sleep in
games that support it. Mute audio, stop frame capture, and dim the power
LED while the lid is closed. Add colorOverride and sleepButtons config
options. Change defaultSave from u16 to u8.

+++ source/arm11/save_type.c
@@ -45,7 +45,7 @@ static u16 checkSaveOverride(const u32 gameCode) // Save type overrides for mode
 	return 0xFF;
 }
 
-u16 detectSaveType(const u32 romSize, const u16 defaultSave)
+u16 detectSaveType(const u32 romSize, const u8 defaultSave)
 {
 	const u32 *romPtr = (u32*)LGY_ROM_LOC;
 	u16 saveType = checkSaveOverride(romPtr[0xAC / 4]);
